Implement StepOne.run()
This implements StepOne.run() such that the unit test passes correctly.

This uses IntStream.rangeClosed() as basic for loops are a source of
off-by-one errors, and this is more explicit about what it's doing.

There is the obvious temptation to implement this as a map operation
followed by a .collect(joining(" ")), but this will require keeping the
entire output in memory before sending it to the output. (We do copy the
output to memory for testing, using a ByteArrayOutputStream wrapped in a
PrintWriter.)