Implement Step 3
This commit implements step 3 such that the unit tests pass.

The use of AtomicInteger as the counters is for their mutable property
and their one-step increment method. As we're running single-threaded,
their thread-safe property is not important, but their lock-free operation
should mean that their performance is not a huge issue.